requirements.txt
# General
black>=23.12.1,<24
python-dotenv>=1.0.0,<2
pytest>=8.1.1,<9

# Flask
Flask>=3.0.0,<4
Flask-SQLAlchemy>=3.1.1,<4
    pymysql>=1.1.0,<2
    cryptography>=41.0.7,<42
Flask-Migrate>=4.0.5,<5
pytest-flask>=1.3.0,<2
flask-restx>=1.3.0,<2
Flask-JSON>=0.4.0,<1
Flask-WTF>=1.2.1,<2
Flask-Login>=0.6.3,<1
Flask-Bcrypt>=1.0.1,<2
